+++ etc/module.xml
@@ -1,5 +1,5 @@
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
-	<module name="Fetchify_Fetchify" setup_version="2.8.1">
+	<module name="Fetchify_Fetchify" setup_version="2.8.2">
 		<sequence>
 			<module name="Magento_Persistent"/>
 		</sequence>
