+++ registration.php
@@ -5,7 +5,7 @@
  * See COPYING.txt for license details.
  */
 \Magento\Framework\Component\ComponentRegistrar::register(
-  \Magento\Framework\Component\ComponentRegistrar::MODULE,
-  'Fetchify_Fetchify',
-  __DIR__
+    \Magento\Framework\Component\ComponentRegistrar::MODULE,
+    'Fetchify_Fetchify',
+    __DIR__
 );
